Cyan: Enable DVFS
This PcdDvfsEnable field was left at 0 earlier since kernel was not
ready for it yet. Changing it to be enabled now, since kernel anyway will keep
it disabled until it is ready for DVFS enabling.
Setting this field in coreboot helps with the Vnn going to around 0.8 in S3

BUG=chrome-os-partner:41825
BRANCH=None
TEST=Boot to OS and also verify if S3 works

Change-Id: Ia2918bf1230b19b0372dcd0574ff468e8bbc3713
Signed-off-by: Hannah Williams <hannah.williams@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/284662
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>