braswell: allow dirty cache line evictions for SMRAM to stick
The BUNIT controls the policy for read/write access to physical
memory. For the SMRAM range the policy was not allowing dirty
evictions to the SMRAM when the core causing the eviction was not
in SMM mode. This could happen when the SMM handler dirtied a line
and then RSM'd back into non-SMM mode. The cache line was dirtied
while in SMM mode, but when that particular cache line was evicted
it would be silently dropped. Fix this by allowing the BUNIT to honor
writes to the SMRAM range while the evicting core is not in SMM mode.
The core SMRR msr provides the mechanism for disallowing general access
to the SMRAM region while it is not in SMM mode.

BUG=chrome-os-partner:43091
BRANCH=firmware-strago-7287.B
TEST=Run suspend_stress_test and ensure there is no hang SMI handler
on suspend-path.
Signed-off-by: Chiranjeevi Rapolu <chiranjeevi.rapolu@intel.com>

Change-Id: I30d97172a0eb536c79d8c299738b3e16835f5b74
Reviewed-on: https://chromium-review.googlesource.com/292791
Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Tested-by: Hannah Williams <hannah.williams@intel.com>
Commit-Queue: Hannah Williams <hannah.williams@intel.com>