Arkham: New color values as per UX team feedback
UX team experimented with Arkham and suggested the new color
values for Arkham. The changes are documented in the "Jetstream
Hardware UI" at http://go/jetstream-hw-ui (Page 5).

Some of the comments were not in sync with the code, which has been
fixed. Though the colors were not named here, I have tried matching it
up with the color values given in the hardware-ui document and change
it for Arkham.

BUG=chrome-os-partner:44798
TEST=When the AP boots up, the colors should be seen at the bottom
of the device, depending on the state of the device. Alternatively,
move the device to different states manually by appropriate actions (
like dev mode, rec mode etc) and observe the colors.
go/jetstream-modes explains how to change the state of device.
BRANCH=None

Change-Id: I7b27c7a740b39d95915a772e02caf6d88cc578ac
Signed-off-by: Suresh Rajashekara <sureshraj@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/296788
Reviewed-by: Toshi Kikuchi <toshik@chromium.org>
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>