Buddy: Add ACPI code to support wake-on-lan
There needs to be an ACPI linkage to provide the power resource
needed to wake this device so the kernel will enable the SCI
before going to suspend.
A link is added for both NIC and WLAN, but it is only tested
on the NIC.
This is a forward port from Duncan's beltino patch.

BUG=None
BRANCH=firmware-auron-6301.B
TEST=build and boot on buddy, suspend and wake with etherwake
Signed-off-by: Ted Kuo <tedkuo@ami.com.tw>

Change-Id: I58d3bf38260d0812067cbe26b224b6dba89ee5f8
Reviewed-on: https://chromium-review.googlesource.com/299615
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>
Commit-Queue: Ted Kuo <tedkuo@ami.com.tw>
Tested-by: Ted Kuo <tedkuo@ami.com.tw>