Arkham: More LED color changes as per UX feedback
UX team experimented with Arkham and suggested the new color
values for Arkham. The changes are documented in the "Jetstream
Hardware UI" at http://go/jetstream-hw-ui (Page 5). The latest
changes are highlighted in blue.

BUG=chrome-os-partner:44798
TEST=When the AP boots up, the colors should be seen at the bottom
of the device, depending on the state of the device. Alternatively,
move the device to different states manually by appropriate actions (
like dev mode, rec mode etc) and observe the colors.
go/jetstream-modes explains how to change the state of device.
BRANCH=None

Change-Id: I3acd6ac6771b45634a2d3b1378e085b8b0e046f5
Signed-off-by: Suresh Rajashekara <sureshraj@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/301392
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>