Cyan :Enable DPTF charger/TSR1/TSR2 participant
With DPTF running we need to enable charger participant.

DPTF keeps polling PPPC every 3 sec and as long as it see the
proper value when it reads PPPC it will enable charging as long
as temperature is not too high.

Requesting to charge the battery on every AC connect query(_Q04)
without knowing if battery expect charging is not good, hence we
need to remove that too.

BUG=chrome-os-partner:45780
BRANCH=firmware-strago-7287.B,firmware-cyan-7287.57.B
TEST=Plug/Unplug AC Adapter multiple times and make sure device is
     charging  properly.

Change-Id: I188e80e6688d0bac5bed6dd64cd2d0feefa30d3f
Signed-off-by: Freddy Paul <freddy.paul@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/304070
Reviewed-by: Aaron Durbin <adurbin@chromium.org>