ultima: Fix for Crossystem "wpsw_cur" status
The GPIO mapping was incorrect for wpsw_cur. The GPIOs for East community were
in two ranges - 0: INT33FF:02 GPIOS [373 - 384] PINS [0 - 11] and
12: INT33FF:02 GPIOS [385 - 396] PINS [15 - 26]
The discontinuity was not accounted for, hence the error.
Original offset was 0x16 whereas it should be 0x13

BUG=chrome-os-partner:47760
BRANCH=None
Signed-off-by: Keith Tzeng <Keith.Tzeng@quantatw.com>
TEST=Run crossystem and test wpsw_cur entry. If screw is present, it should be 1 and if not present,
it should be 0

Change-Id: Ic0bbc2929d073b63d596703f4d06db608db8625b
Reviewed-on: https://chromium-review.googlesource.com/322971
Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Commit-Queue: Keith Tzeng <keith.tzeng@quantatw.com>
Tested-by: Keith Tzeng <keith.tzeng@quantatw.com>