chromeos: Add vbnv wrapper for the different backends
Add a wrapper around the vbnv implementations and call into the different
backend functions from there.  Also move some of the common functions to
the common code and simplify the backend drivers.  This will allow some
of the code to be re-used so the CMOS backend can backup the data into
the flash backend.

One side effect of this is that the cache of VBNV was removed from CMOS
and EC backends and moved into the VBNV wrapper, but the flash backend
also still has a separate cache because it has more state and complexity
in the implementation.  The wrapper cached data is not used for normal
vbnv_read/vbnv_write because some callers need the ability to force a
write if the backend storage is cleared (i.e. CMOS clear).

BUG=chrome-os-partner:47915
BRANCH=glados
TEST=build and boot on chell

Change-Id: Ia97f6607c5ad837b9aa10b45211137221ccb93a0
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Previous-Reviewed-on: https://chromium-review.googlesource.com/324120
(cherry picked from commit 691fb5db0b27aee20b1a2b87e5f45c9ab3d1c87e)
Reviewed-on: https://chromium-review.googlesource.com/324955