skylake: Check for power failure when WAK_STS is not set
The PCH does not set PM1_STS[WAK_STS] bit when waking from a
G3 state, which is triggered by hiberante now on chell when we
do a PMIC shutdown.  This means the checks for S5 wake are not
done and instead it is logged as a wake from S0.

BUG=chrome-os-partner:50076
BRANCH=glados
TEST=pass firmware_EventLog test on chell

Change-Id: I4472498468d620fe69f2b68710e818a4ad287382
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Previous-Reviewed-on: https://chromium-review.googlesource.com/326888
(cherry picked from commit 821fafe4fc3aece7320623c46d4f29a299e1e8d8)
Reviewed-on: https://chromium-review.googlesource.com/327334