UPSTREAM: soc/*: fix uart's regwidth specification in cbtables
coreboot passes information about the serial port implementation to
payloads through a cbtables entry.
We set the register width to 1 on most SoCs because that looked as good
a default as any, but checking the uart structs they use, it's 4 for all
of them.

Original-Change-Id: I9848f79737106dc32f864ca901c0bc48f489e6b8
Original-Signed-off-by: Patrick Georgi <pgeorgi@chromium.org>
Original-Reviewed-on: https://review.coreboot.org/13746
Original-Reviewed-by: Julius Werner <jwerner@chromium.org>

Change-Id: I6794aba1fd4b883b67032427c48fe92d4d356804
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Reviewed-on: https://chromium-review.googlesource.com/329544
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>